Initial commit of test i2c utils for ALPACA infrastructure
* alpaca_i2c_utils.h - definition of i2c device and enum of attached i2c devices
* alpaca_i2c_utils.c - Implementation of i2c write, read, and read with repeated
  start (read after write) and simple main testing the functionality by
  read/writing from/to eeprom and the idt8a34001
* idt8a34001_regs.c/h - definition of the registers for programming the idt8a34001

These should be general purpose enough that potentially could merge support for
HTG LMK/LMX clocking. Could also do the same for teh CLK104 board. However, that
board has the gpio expander that would also need to be toggled which is not
specifically i2c releated and so if to include CLK104 would be a hierarchy
dependency.