add i2c0 file descriptor and identify i2c-mux problem, add proper i2c bus close
Making an i2c driver that is extendable has its problems when needing to be
versatile to different scenarios and configurations. In this context is the i2c
coniguration for the rfsoc boards where the mpsoc has i2c buses each with mux to
extend attached i2c devices. For the rf clocking htg uses i2c0 and xilinx eval
boards use i2c1. The inital code only targeted i2c1 and now really need both.

Not just in the context of the rf clocks either, but all board i2c peripherals.
So it makes sense to just always instance both since generally both busses are
configured in mpsoc boards (although not guaranteed).

So staging these commits before I start hacking and testing a way to determine
parent device. My first thought is to just have a parent fd object in my struct
but I am not sure if this will cause problems when a device doesn't sit on a
mux.