Attempt to consolodate platform definitions for rfplls
(also updates to continue to add i2c infrastructure and pynq2x2 definitions)

Take all of the rfpll platform specific definitions from the test standalone and
make a single header and implementations that builds definitions based on the
rfsoc platform definition. A new platform header is used to be included for the
different hardware peripherals (first i2c and now lmk/lmx plls).

The implementation in the test files has not changed, just pulling out what can
be consolodated. From here would be to continue to make clk programming methods.