addition of rfsoc4x2 platform and spi
tested to at least program the rfsoc4x2 and not break the zcu216. Read back is
not currently working on rfsoc4x2 but seems to still be working on zcu216

rfsoc4x2 uses spi directly while the other rfsoc boards have used an i2c to spi
bridge. So this is a brute force hack to just get spi in there.

might be worthwhile to rethink the implementation since it is becoming unweildy
to have the different cases based on platform